change: start to rearrange screens
